pyproject.toml
[project]
name = "financial-ai-copilot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gradio>=6.8.0",
    "langchain-mcp-adapters>=0.2.1",
    "langchain-mongodb>=0.11.0",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-mongodb>=0.3.1",
    "langgraph-store-mongodb>=0.2.0",
    "langgraph-prebuilt>=1.0.8",
    "nodejs>=0.1.1",
    "openai>=2.24.0",
    "pymongo>=4.15.5",
    "streamlit>=1.54.0",
    "langchain-google-genai>=4.2.1",
    "yfinance>=1.2.0",
]
